update integer cpf in migrate user

+++ database/migrations/2021_10_02_000231_create_table_users.php
@@ -17,7 +17,7 @@ class CreateTableUsers extends Migration
 
                 $table->string('name');
 
-                $table->integer('cpf')
+                $table->bigInteger('cpf')
                     ->unique();
 
                 $table->string('email')
